+++ src/main/java/com/uwetrottmann/trakt5/entities/SyncEpisode.java
@@ -35,7 +35,7 @@ public class SyncEpisode {
 
     public OffsetDateTime collected_at;
     /**
-     * Warning: Trakt is planning to only store and return minute-precision timestamps for watched_at. So seconds and
+     * Warning: Trakt only stores and returns minute-precision timestamps for the watched at time. So seconds and
      * nanoseconds will always be zero. Using {@code .truncatedTo(ChronoUnit.MINUTES)} can be helpful.
      */
     public OffsetDateTime watched_at;
@@ -73,6 +73,10 @@ public class SyncEpisode {
         return this;
     }
 
+    /**
+     * Warning: Trakt only stores and returns minute-precision timestamps for the watched at time. So seconds and
+     * nanoseconds will always be zero. Using {@code .truncatedTo(ChronoUnit.MINUTES)} can be helpful.
+     */
     @Nonnull
     public SyncEpisode watchedAt(OffsetDateTime watchedAt) {
         this.watched_at = watchedAt;