+++ src/main/java/com/uwetrottmann/trakt5/entities/SyncMovie.java
@@ -33,7 +33,7 @@ public class SyncMovie {
 
     public OffsetDateTime collected_at;
     /**
-     * Warning: Trakt is planning to only store and return minute-precision timestamps for watched_at. So seconds and
+     * Warning: Trakt only stores and returns minute-precision timestamps for the watched at time. So seconds and
      * nanoseconds will always be zero. Using {@code .truncatedTo(ChronoUnit.MINUTES)} can be helpful.
      */
     public OffsetDateTime watched_at;
@@ -59,6 +59,10 @@ public class SyncMovie {
         return this;
     }
 
+    /**
+     * Warning: Trakt only stores and returns minute-precision timestamps for the watched at time. So seconds and
+     * nanoseconds will always be zero. Using {@code .truncatedTo(ChronoUnit.MINUTES)} can be helpful.
+     */
     @Nonnull
     public SyncMovie watchedAt(OffsetDateTime watchedAt) {
         this.watched_at = watchedAt;